refactor: move profile fields to notes page
Remove ProfileSection component from about page and
integrate description and personality fields directly
into notes page with auto-save on blur functionality.
Update documentation to reflect completed notes page
implementation with all profile fields now editable
in a single dedicated location.
